refactor: clean up Anthropic provider internals
Inline constants, consolidate factory/register into index.ts, and
extract duplicated API request building into buildRequestParams().
Eliminates 3 files (-257 lines, +79 lines).
