+++ src/providers/anthropic/index.ts
@@ -7,32 +7,17 @@
 /**
  * Anthropic (Claude) provider module.
  *
- * Provides direct access to Anthropic's Messages API for Claude models.
- *
  * @module providers/anthropic
- *
- * @example
- * ```typescript
- * import { Anthropic, registerAnthropic } from "adk-llm-bridge";
- *
- * // Option 1: Factory function
- * const llm = Anthropic("claude-sonnet-4-5-20250929");
- *
- * // Option 2: Registry integration
- * registerAnthropic();
- * const agent = new LlmAgent({ model: "claude-sonnet-4-5-20250929" });
- * ```
  */
 
-export { AnthropicLlm } from "./anthropic-llm";
-export {
-  ANTHROPIC_BASE_URL,
-  ANTHROPIC_ENV,
-  ANTHROPIC_MODEL_PATTERNS,
-  DEFAULT_ANTHROPIC_MAX_TOKENS,
-} from "./constants";
+import { LLMRegistry } from "@google/adk";
+import { resetProviderConfig, setProviderConfig } from "../../config";
+import type { AnthropicProviderConfig, AnthropicRegisterOptions } from "../../types";
+import { AnthropicLlm } from "./anthropic-llm";
+
+// Re-exports
+export { AnthropicLlm, ANTHROPIC_MODEL_PATTERNS } from "./anthropic-llm";
 export type { ConvertedAnthropicRequest } from "./converters/request";
-// Export converters for potential custom implementations
 export { convertAnthropicRequest } from "./converters/request";
 export type {
   AnthropicStreamAccumulator,
@@ -43,9 +28,42 @@ export {
   convertAnthropicStreamEvent,
   createAnthropicStreamAccumulator,
 } from "./converters/response";
-export { Anthropic } from "./factory";
-export {
-  _resetAnthropicRegistration,
-  isAnthropicRegistered,
-  registerAnthropic,
-} from "./register";
+
+/**
+ * Creates an Anthropic (Claude) LLM instance.
+ *
+ * @param model - The Claude model to use
+ * @param options - Optional configuration options
+ * @returns A configured Anthropic LLM instance
+ */
+export function Anthropic(
+  model: string,
+  options?: Omit<AnthropicProviderConfig, "model">,
+): AnthropicLlm {
+  return new AnthropicLlm({ model, ...options });
+}
+
+// --- Registration (singleton pattern) ---
+
+let registered = false;
+
+export function registerAnthropic(options?: AnthropicRegisterOptions): void {
+  if (registered) {
+    console.warn("[adk-llm-bridge] anthropic already registered, skipping");
+    return;
+  }
+  if (options) {
+    setProviderConfig("anthropic", options);
+  }
+  LLMRegistry.register(AnthropicLlm);
+  registered = true;
+}
+
+export function isAnthropicRegistered(): boolean {
+  return registered;
+}
+
+export function _resetAnthropicRegistration(): void {
+  registered = false;
+  resetProviderConfig("anthropic");
+}
